Init repo
A skeleton of the repo.

We have `minesweeper.js` which is the entry point for the application.
Run `node minesweeper.js` and then visit localhost:3000.

node version tested on: `20.11.1`

The only thing which renders is a hello world from the index.html file
and a console log which will appear in the console browser.